Query webOS API version instead of kernel version
When a user queries the com.palm.properties.version property he wants to get the webOS API
version and not the version of the kernel the system is running with. This is what legacy
did and still some components in the system expect to get returned.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>